Category transactions batch updates
