+++ packages/desktop-client/src/components/mobile/accounts/AccountTransactions.jsx
@@ -24,29 +24,15 @@ import {
   useDefaultSchedulesQueryTransform,
 } from 'loot-core/client/data-hooks/schedules';
 import * as queries from 'loot-core/client/queries';
-import { pagedQuery, runQuery } from 'loot-core/client/query-helpers';
+import { pagedQuery } from 'loot-core/client/query-helpers';
 import { listen, send } from 'loot-core/platform/client/fetch';
-import * as monthUtils from 'loot-core/shared/months';
-import { q } from 'loot-core/shared/query';
-import {
-  deleteTransaction,
-  isPreviewId,
-  realizeTempTransactions,
-  ungroupTransaction,
-  ungroupTransactions,
-  updateTransaction,
-} from 'loot-core/shared/transactions';
-import { applyChanges, integerToCurrency } from 'loot-core/shared/util';
+import { isPreviewId } from 'loot-core/shared/transactions';
 
-import { useAccounts } from '../../../hooks/useAccounts';
-import { useCategories } from '../../../hooks/useCategories';
 import { useDateFormat } from '../../../hooks/useDateFormat';
 import { useNavigate } from '../../../hooks/useNavigate';
-import { usePayees } from '../../../hooks/usePayees';
 import { usePreviewTransactions } from '../../../hooks/usePreviewTransactions';
 import { SelectedProvider, useSelected } from '../../../hooks/useSelected';
 import { useSyncedPref } from '../../../hooks/useSyncedPref';
-import { useUndo } from '../../../hooks/useUndo';
 import { styles, theme } from '../../../style';
 import { Text } from '../../common/Text';
 import { View } from '../../common/View';
@@ -152,9 +138,6 @@ function AccountName({ account, pending, failed }) {
 }
 
 function TransactionListWithPreviews({ account }) {
-  const accounts = useAccounts();
-  const payees = usePayees();
-  const { list: categories } = useCategories();
   const [currentQuery, setCurrentQuery] = useState();
   const [isSearching, setIsSearching] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
@@ -283,365 +266,7 @@ function TransactionListWithPreviews({ account }) {
   const balanceCleared = queries.accountBalanceCleared(account);
   const balanceUncleared = queries.accountBalanceUncleared(account);
 
-  // Note: deleted transactions will still be in this list so that
-  // when the user undoes the delete, the transactions will still be selected.
   const selectedInst = useSelected('transactions', allTransactions);
-  const selectedTransactions = selectedInst.items;
-  const hasMoreThanOneSelected = selectedTransactions.size > 1;
-
-  const { showUndoNotification } = useUndo();
-
-  const onBatchEdit = async (name, ids) => {
-    const { data } = await runQuery(
-      q('transactions')
-        .filter({ id: { $oneof: ids } })
-        .select('*')
-        .options({ splits: 'grouped' }),
-    );
-    const transactions = ungroupTransactions(data);
-
-    const onChange = async (name, value, mode) => {
-      let transactionsToChange = transactions;
-
-      const newValue = value === null ? '' : value;
-      const changes = { deleted: [], updated: [] };
-
-      // Cleared is a special case right now
-      if (name === 'cleared') {
-        // Clear them if any are uncleared, otherwise unclear them
-        value = !!transactionsToChange.find(t => !t.cleared);
-      }
-
-      const idSet = new Set(ids);
-
-      transactionsToChange.forEach(trans => {
-        if (name === 'cleared' && trans.reconciled) {
-          // Skip transactions that are reconciled. Don't want to set them as
-          // uncleared.
-          return;
-        }
-
-        if (!idSet.has(trans.id)) {
-          // Skip transactions which aren't actually selected, since the query
-          // above also retrieves the siblings & parent of any selected splits.
-          return;
-        }
-
-        if (name === 'notes') {
-          if (mode === 'prepend') {
-            value =
-              trans.notes === null ? newValue : newValue + ' ' + trans.notes;
-          } else if (mode === 'append') {
-            value =
-              trans.notes === null ? newValue : trans.notes + ' ' + newValue;
-          } else if (mode === 'replace') {
-            value = newValue;
-          }
-        }
-        const transaction = {
-          ...trans,
-          [name]: value,
-        };
-
-        if (name === 'account' && trans.account !== value) {
-          transaction.reconciled = false;
-        }
-
-        const { diff } = updateTransaction(transactionsToChange, transaction);
-
-        // TODO: We need to keep an updated list of transactions so
-        // the logic in `updateTransaction`, particularly about
-        // updating split transactions, works. This isn't ideal and we
-        // should figure something else out
-        transactionsToChange = applyChanges(diff, transactionsToChange);
-
-        changes.deleted = changes.deleted
-          ? changes.deleted.concat(diff.deleted)
-          : diff.deleted;
-        changes.updated = changes.updated
-          ? changes.updated.concat(diff.updated)
-          : diff.updated;
-        changes.added = changes.added
-          ? changes.added.concat(diff.added)
-          : diff.added;
-      });
-
-      await send('transactions-batch-update', changes);
-
-      let displayValue = value;
-      switch (name) {
-        case 'account':
-          displayValue = accounts.find(a => a.id === value).name;
-          break;
-        case 'category':
-          displayValue = categories.find(c => c.id === value).name;
-          break;
-        case 'payee':
-          displayValue = payees.find(p => p.id === value).name;
-          break;
-        case 'amount':
-          displayValue = integerToCurrency(value);
-          break;
-        default:
-          displayValue = value;
-          break;
-      }
-
-      showUndoNotification({
-        title: 'Batch operation complete',
-        message: `Successfully updated ${name} of ${selectedTransactions.size} transaction${hasMoreThanOneSelected ? 's' : ''} to [${displayValue}](#${displayValue}).`,
-        messageActions: {
-          [displayValue]: () => {
-            switch (name) {
-              case 'account':
-                navigate(`/accounts/${value}`);
-                break;
-              case 'category':
-                navigate(`/categories/${value}`);
-                break;
-              case 'payee':
-                navigate(`/payees`);
-                break;
-              default:
-                break;
-            }
-          },
-        },
-      });
-    };
-
-    const pushPayeeAutocompleteModal = () => {
-      dispatch(
-        pushModal('payee-autocomplete', {
-          onSelect: payeeId => onChange(name, payeeId),
-        }),
-      );
-    };
-
-    const pushAccountAutocompleteModal = () => {
-      dispatch(
-        pushModal('account-autocomplete', {
-          onSelect: accountId => onChange(name, accountId),
-        }),
-      );
-    };
-
-    const pushEditField = () => {
-      dispatch(pushModal('edit-field', { name, onSubmit: onChange }));
-    };
-
-    const pushCategoryAutocompleteModal = () => {
-      // Only show balances when all selected transaction are in the same month.
-      const transactionMonth = transactions[0]?.date
-        ? monthUtils.monthFromDate(transactions[0]?.date)
-        : null;
-      const transactionsHaveSameMonth =
-        transactionMonth &&
-        transactions.every(
-          t => monthUtils.monthFromDate(t.date) === transactionMonth,
-        );
-      dispatch(
-        pushModal('category-autocomplete', {
-          month: transactionsHaveSameMonth ? transactionMonth : undefined,
-          onSelect: categoryId => onChange(name, categoryId),
-        }),
-      );
-    };
-
-    if (
-      name === 'amount' ||
-      name === 'payee' ||
-      name === 'account' ||
-      name === 'date'
-    ) {
-      const reconciledTransactions = transactions.filter(t => t.reconciled);
-      if (reconciledTransactions.length > 0) {
-        dispatch(
-          pushModal('confirm-transaction-edit', {
-            onConfirm: () => {
-              if (name === 'payee') {
-                pushPayeeAutocompleteModal();
-              } else if (name === 'account') {
-                pushAccountAutocompleteModal();
-              } else {
-                pushEditField();
-              }
-            },
-            confirmReason: 'batchEditWithReconciled',
-          }),
-        );
-        return;
-      }
-    }
-
-    if (name === 'cleared') {
-      // Cleared just toggles it on/off and it depends on the data
-      // loaded. Need to clean this up in the future.
-      onChange('cleared', null);
-    } else if (name === 'category') {
-      pushCategoryAutocompleteModal();
-    } else if (name === 'payee') {
-      pushPayeeAutocompleteModal();
-    } else if (name === 'account') {
-      pushAccountAutocompleteModal();
-    } else {
-      pushEditField();
-    }
-  };
-
-  const onBatchDuplicate = async ids => {
-    const onConfirmDuplicate = async ids => {
-      const { data } = await runQuery(
-        q('transactions')
-          .filter({ id: { $oneof: ids } })
-          .select('*')
-          .options({ splits: 'grouped' }),
-      );
-
-      const changes = {
-        added: data
-          .reduce((newTransactions, trans) => {
-            return newTransactions.concat(
-              realizeTempTransactions(ungroupTransaction(trans)),
-            );
-          }, [])
-          .map(({ sort_order, ...trans }) => ({ ...trans })),
-      };
-
-      await send('transactions-batch-update', changes);
-
-      showUndoNotification({
-        title: 'Batch operation complete',
-        message: `Successfully duplicated ${selectedTransactions.size} transaction${hasMoreThanOneSelected ? 's' : ''}.`,
-      });
-    };
-
-    await checkForReconciledTransactions(
-      ids,
-      'batchDuplicateWithReconciled',
-      onConfirmDuplicate,
-    );
-  };
-
-  const onBatchDelete = async ids => {
-    const onConfirmDelete = ids => {
-      dispatch(
-        pushModal('confirm-transaction-delete', {
-          message: hasMoreThanOneSelected
-            ? `Are you sure you want to delete these ${ids.length} transaction${hasMoreThanOneSelected ? 's' : ''}?`
-            : undefined,
-          onConfirm: async () => {
-            const { data } = await runQuery(
-              q('transactions')
-                .filter({ id: { $oneof: ids } })
-                .select('*')
-                .options({ splits: 'grouped' }),
-            );
-            let transactions = ungroupTransactions(data);
-
-            const idSet = new Set(ids);
-            const changes = { deleted: [], updated: [] };
-
-            transactions.forEach(trans => {
-              const parentId = trans.parent_id;
-
-              // First, check if we're actually deleting this transaction by
-              // checking `idSet`. Then, we don't need to do anything if it's
-              // a child transaction and the parent is already being deleted
-              if (!idSet.has(trans.id) || (parentId && idSet.has(parentId))) {
-                return;
-              }
-
-              const { diff } = deleteTransaction(transactions, trans.id);
-
-              // TODO: We need to keep an updated list of transactions so
-              // the logic in `updateTransaction`, particularly about
-              // updating split transactions, works. This isn't ideal and we
-              // should figure something else out
-              transactions = applyChanges(diff, transactions);
-
-              changes.deleted = diff.deleted
-                ? changes.deleted.concat(diff.deleted)
-                : diff.deleted;
-              changes.updated = diff.updated
-                ? changes.updated.concat(diff.updated)
-                : diff.updated;
-            });
-
-            await send('transactions-batch-update', changes);
-            showUndoNotification({
-              type: 'warning',
-              title: 'Batch operation complete',
-              message: `Successfully deleted ${selectedTransactions.size} transaction${hasMoreThanOneSelected ? 's' : ''}.`,
-            });
-          },
-        }),
-      );
-    };
-
-    await checkForReconciledTransactions(
-      ids,
-      'batchDeleteWithReconciled',
-      onConfirmDelete,
-    );
-  };
-
-  const onLinkSchedule = ids => {
-    dispatch(
-      pushModal('schedule-link', {
-        transactionIds: ids,
-        getTransaction: id => transactions.find(t => t.id === id),
-        onScheduleLinked: schedule => {
-          // TODO: When schedule becomes available in mobile, update undo notification message
-          // with `messageActions` to open the schedule when the schedule name is clicked.
-          showUndoNotification({
-            title: 'Batch operation complete',
-            message: `Successfully linked ${selectedTransactions.size} transaction${hasMoreThanOneSelected ? 's' : ''} to ${schedule.name}.`,
-          });
-        },
-      }),
-    );
-  };
-
-  const onUnlinkSchedule = async ids => {
-    await send('transactions-batch-update', {
-      updated: ids.map(id => ({ id, schedule: null })),
-    });
-    showUndoNotification({
-      title: 'Batch operation complete',
-      message: `Successfully unlinked ${selectedTransactions.size} transaction${hasMoreThanOneSelected ? 's' : ''} from their respective schedules.`,
-    });
-  };
-
-  const onSetTransfer = () => {
-    // Add support when All accounts/For budget/Off budget views are added.
-  };
-
-  const checkForReconciledTransactions = async (
-    ids,
-    confirmReason,
-    onConfirm,
-  ) => {
-    const { data } = await runQuery(
-      q('transactions')
-        .filter({ id: { $oneof: ids }, reconciled: true })
-        .select('*')
-        .options({ splits: 'grouped' }),
-    );
-    const transactions = ungroupTransactions(data);
-    if (transactions.length > 0) {
-      dispatch(
-        pushModal('confirm-transaction-edit', {
-          onConfirm: () => {
-            onConfirm(ids);
-          },
-          confirmReason,
-        }),
-      );
-    } else {
-      onConfirm(ids);
-    }
-  };
 
   return (
     <SelectedProvider instance={selectedInst}>
@@ -656,12 +281,6 @@ function TransactionListWithPreviews({ account }) {
         onSearch={onSearch}
         onOpenTransaction={onOpenTransaction}
         onRefresh={onRefresh}
-        onBatchEdit={onBatchEdit}
-        onBatchDuplicate={onBatchDuplicate}
-        onSetTransfer={onSetTransfer}
-        onLinkSchedule={onLinkSchedule}
-        onUnlinkSchedule={onUnlinkSchedule}
-        onBatchDelete={onBatchDelete}
       />
     </SelectedProvider>
   );
